requirements.txt
bittensor==6.9.3
diffusers==0.25.0
transformers==4.39.1
accelerate==0.27.2
cryptography==42.0.0
ImageHash==4.3.1
loguru==0.7.0
Pillow==10.2.0
pydantic==1.10.13
PyYAML==6.0.1
Requests==2.31.0
setuptools==68.0.0
slowapi==0.1.8
tqdm==4.65.0
omegaconf==2.3.0
gitpython==3.1.40
httpx==0.26.0
ray==2.9.2
ray[serve]
controlnet_aux==0.0.7
discord-webhook
timm
wandb
pandas